fix: redirect disabled timesheets route directly
