+++ apps/web/src/routes/_protected.workspaces/$workspaceId/index.tsx
@@ -1,31 +1,12 @@
-import { createFileRoute, redirect } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 
-import { ensureRouteQueryData } from "@/lib/react-query";
-import { viewsOptions } from "@/routes/_protected.workspaces/$workspaceId/-queries/views";
+import { redirectToDefaultWorkspaceView } from "@/routes/_protected.workspaces/$workspaceId/-default-view-redirect";
 
 export const Route = createFileRoute("/_protected/workspaces/$workspaceId/")({
   beforeLoad: async ({ context, params }) => {
-    const qc = context.queryClient;
-    const opts = viewsOptions(params.workspaceId);
-
-    // Invalidate first to avoid serving stale cache from a
-    // previous workspace that had no views.
-    await qc.invalidateQueries({ queryKey: opts.queryKey });
-
-    const views = await ensureRouteQueryData(qc, opts);
-
-    const firstView = views.at(0);
-    if (!firstView) {
-      throw redirect({ to: "/workspaces", replace: true });
-    }
-
-    throw redirect({
-      to: "/workspaces/$workspaceId/$viewId",
-      params: {
-        workspaceId: params.workspaceId,
-        viewId: firstView.id,
-      },
-      replace: true,
+    await redirectToDefaultWorkspaceView({
+      queryClient: context.queryClient,
+      workspaceId: params.workspaceId,
     });
   },
 });
