+++ apps/web/src/routes/_protected.workspaces/$workspaceId/timesheets.tsx
@@ -1,4 +1,6 @@
-import { createFileRoute, redirect } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
+
+import { redirectToDefaultWorkspaceView } from "@/routes/_protected.workspaces/$workspaceId/-default-view-redirect";
 
 export const Route = createFileRoute(
   "/_protected/workspaces/$workspaceId/timesheets",
@@ -6,13 +8,10 @@ export const Route = createFileRoute(
   // Time tracking is intentionally excluded from the current product surface.
   // Keep this route redirect-only so abandoned /timesheets navigations cannot
   // mount the old billing query tree before the redirect commits.
-  beforeLoad: ({ params }) => {
-    throw redirect({
-      to: "/workspaces/$workspaceId",
-      params: {
-        workspaceId: params.workspaceId,
-      },
-      replace: true,
+  beforeLoad: async ({ context, params }) => {
+    await redirectToDefaultWorkspaceView({
+      queryClient: context.queryClient,
+      workspaceId: params.workspaceId,
     });
   },
 });
